ch08_retrieval_quality/requirements.txt
numpy==2.2.6
rich==14.2.0
faiss-cpu==1.13.2
pytest==8.3.5
